Fix timestamp conversion in katcp testutils test client
git-svn-id: https://katfs.kat.ac.za/svnDS/code/katcp-python/branches/katcpv5@15923 b56c7a7d-ac47-0410-a905-b6f7de78896a

+++ katcp/testutils.py
@@ -241,10 +241,10 @@ class BlockingTestClient(client.BlockingClient):
                 (status, sensorname))
 
         try:
-            timestamp = int(timestamp)
+            timestamp = float(timestamp)
         except ValueError, e:
             self.test.fail("Could not convert timestamp %r of sensor '%s' to"
-                           " type %r: %s" % (timestamp, sensorname, int, e))
+                           " type %r: %s" % (timestamp, sensorname, float, e))
 
         return value, status, timestamp
